requirements.txt
customtkinter>=5.2.0
pyodbc>=5.0.0
pandas>=2.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tkcalendar>=1.6.0

